Move bad request check to ErrorLogger
This way they also do not get logged if they occur through a supervisor/resque process
